requirements.txt
openai-whisper
sounddevice
soundfile
numpy
torch
librosa